db-sources: ContainerFactory.Singleton -> ContainerFactoryWrapper
