+++ airbyte-cdk/java/airbyte-cdk/db-sources/src/testFixtures/java/io/airbyte/cdk/testutils/ContainerFactoryWrapper.java
@@ -1,0 +1,85 @@
+/*
+ * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
+ */
+
+package io.airbyte.cdk.testutils;
+
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.concurrent.ConcurrentHashMap;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.testcontainers.containers.JdbcDatabaseContainer;
+import org.testcontainers.containers.output.Slf4jLogConsumer;
+import org.testcontainers.utility.DockerImageName;
+
+/**
+ * This class wraps a specific shared testcontainer instance, which is created exactly once.
+ */
+class ContainerFactoryWrapper<C extends JdbcDatabaseContainer<?>> {
+
+  static private final Logger LOGGER = LoggerFactory.getLogger(ContainerFactoryWrapper.class);
+  static private final ConcurrentHashMap<String, ContainerFactoryWrapper<?>> LAZY = new ConcurrentHashMap<>();
+
+  @SuppressWarnings("unchecked")
+  static <C extends JdbcDatabaseContainer<?>> C getOrCreate(String mapKey, ContainerFactory<C> factory) {
+    final ContainerFactoryWrapper<?> singleton = LAZY.computeIfAbsent(mapKey, ContainerFactoryWrapper<C>::new);
+    return ((ContainerFactoryWrapper<C>) singleton).getOrCreate(factory);
+  }
+
+  final String imageName;
+  final List<String> methodNames;
+
+  private C sharedContainer;
+  private RuntimeException containerCreationError;
+
+  private ContainerFactoryWrapper(String imageNamePlusMethods) {
+    final String[] parts = imageNamePlusMethods.split("\\+");
+    this.imageName = parts[0];
+    this.methodNames = Arrays.stream(parts).skip(2).toList();
+  }
+
+  private synchronized C getOrCreate(ContainerFactory<C> factory) {
+    if (sharedContainer == null && containerCreationError == null) {
+      try {
+        create(imageName, factory, methodNames);
+      } catch (RuntimeException e) {
+        sharedContainer = null;
+        containerCreationError = e;
+      }
+    }
+    if (containerCreationError != null) {
+      throw new RuntimeException(
+          "Error during container creation for imageName=" + imageName
+              + ", factory=" + factory.getClass().getName()
+              + ", methods=" + methodNames,
+          containerCreationError);
+    }
+    return sharedContainer;
+  }
+
+  private void create(String imageName, ContainerFactory<C> factory, List<String> methodNames) {
+    LOGGER.info("Creating new shared container based on {} with {}.", imageName, methodNames);
+    try {
+      final var parsed = DockerImageName.parse(imageName);
+      final var methods = new ArrayList<Method>();
+      for (String methodName : methodNames) {
+        methods.add(factory.getClass().getMethod(methodName, factory.getContainerClass()));
+      }
+      sharedContainer = factory.createNewContainer(parsed);
+      sharedContainer.withLogConsumer(new Slf4jLogConsumer(LOGGER));
+      for (Method method : methods) {
+        LOGGER.info("Calling {} in {} on new shared container based on {}.",
+            method.getName(), factory.getClass().getName(), imageName);
+        method.invoke(factory, sharedContainer);
+      }
+      sharedContainer.start();
+    } catch (IllegalAccessException | InvocationTargetException | NoSuchMethodException e) {
+      throw new RuntimeException(e);
+    }
+  }
+
+}
